+++ src/Models/User.php
@@ -46,7 +46,6 @@ class User extends Authenticatable
         'email',
         'password',
         'is_active',
-        'email_verified_at',
     ];
 
     /**