+++ tests/Unit/Enums/LocaleTest.php
@@ -1,0 +1,16 @@
+<?php
+
+use Wave8\Factotum\Base\Enums\Locale;
+
+describe('LocaleEnum', function () {
+    it('check LocaleEnum has exact values', function () {
+
+        $values = array_flip(Locale::getValues()->toArray());
+
+        expect(array_keys($values))->toBe([
+            'it',
+            'en',
+        ]);
+
+    });
+});